rename userManage folder/updated imports to fix webpack compilier warning
